Fix sorting collection
Override updated_date from request because the field in the client is not editable. The reason why the sorting is broken because there's field in edit/create form named "updated_date". And formated to "YYYY-MM-DD". When the form got submitted. The time saved to database is always midnight because the format receive by the API is only date. no time.